FEAT: Add remember me functionality and update country data
Adds a "remember me" feature to the login process. If a user has previously selected "remember me", they will be automatically logged in and navigated to the main activity upon app launch. This is implemented by checking the `isUserRememberMe` status in `SharedPrefManager` within `LoginActivity`.

Also updates the country and city data in `RegisterViewModel`:
- Replaces "USA" with "PS" (Palestine) and its corresponding cities ("Nablus", "Tulkarem", "Ramallah").
- Updates the country code for "PS" to "+970".