TEST: Refactor FavoriteDaoTest and add log mocking to LoginViewModelTest
Refactors `FavoriteDaoTest` to correctly retrieve a list of favorites when calling `getFavoriteByEmail` and asserts against the first item in the list. Also updates the delete test to ensure the favorite is actually deleted.

Adds mocking of the Android `Log` class to `LoginViewModelTest` to prevent `java.lang.RuntimeException: Method d in android.util.Log not mocked` errors during test execution. This is achieved by creating a `LogUtils` class with a `mockLog()` method that returns a `MockedStatic<Log>` instance, which is then closed in the `tearDown()` method of the test.