FEAT: Refactor user registration and password hashing
Refactors the user registration process in `RegisterViewModel` to utilize the `User` model's constructor and setters for input validation. This centralizes validation logic within the `User` model.

Password hashing is now performed within the `UserRepository`'s `insertUser` method and the `User` model's `setPasswordWithHash` method, ensuring passwords are consistently hashed before persistence. The `Hashing.createPasswordHash` method is used for this purpose.

The `User` model constructor and `UserMapper` have been updated to include the `gender` field directly. Validation for last name in `User` model is corrected.

The `RegisterViewModel` now directly attempts to create a `User` object. If successful and the email is not already registered, the user is inserted. Validation exceptions from the `User` model are caught and displayed to the user. Redundant validation methods in `RegisterViewModel` have been removed.

The default country in `RegisterViewModel` has been changed from "USA" to "PS".