REFACTOR: Modify repository to return domain models instead of entities
Updates the `PropertyRepository` and `PropertyRepositoryImpl` to return `LiveData` of domain `Property` objects instead of `PropertyEntity` objects.

Introduces a `toDomainList` method in `PropertyMapper` to convert a list of `PropertyEntity` objects to a list of `Property` domain objects.

Uses `Transformations.map` in `PropertyRepositoryImpl` to apply the mapping from entity to domain model for `getAllProperties` and `getPropertyById` methods.

Also, simplifies the `onSuccess` callback in `savePropertiesToLocalDB` by removing the `null` argument.